% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{metabarplot-colors}
\alias{metabarplot-colors}
\alias{cols1}
\alias{cols_brewerplus}
\alias{cols_kelly}
\title{Color Palettes for Taxonomic Barplots}
\format{
A character vector of hex colors.

A character vector of hex colors.

A character vector of hex colors.
}
\usage{
cols1

cols_brewerplus

cols_kelly
}
\description{
Predefined color palettes for use with \code{\link[=pbar]{pbar()}}.

Originally based on \code{microViz::distinct_palette(pal = "brewerPlus", add = NA)}.

Based on \code{microViz::distinct_palette(pal = "kelly", add = NA)} with gray
colors removed.
}
\keyword{datasets}
