% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabarplot-package.R
\docType{package}
\name{metabarplot}
\alias{metabarplot-package}
\alias{metabarplot}
\title{metabarplot: Taxonomic Barplots for Metabarcoding}
\description{
Utilities to summarize taxonomic abundance and generate stacked barplots.
}
\author{
\strong{Maintainer}: Jelmer Poelstra \email{poelstra.1@osu.edu}

}
\keyword{internal}
